Move InvocationConvention to function package
