+++ presto-main/src/main/java/com/facebook/presto/operator/annotations/FunctionImplementationDependency.java
@@ -13,7 +13,7 @@
  */
 package com.facebook.presto.operator.annotations;
 
-import com.facebook.presto.spi.InvocationConvention;
+import com.facebook.presto.spi.function.InvocationConvention;
 import com.facebook.presto.spi.type.TypeSignature;
 
 import java.util.List;
