+++ presto-main/src/main/java/com/facebook/presto/operator/annotations/ImplementationDependency.java
@@ -15,9 +15,9 @@ package com.facebook.presto.operator.annotations;
 
 import com.facebook.presto.metadata.BoundVariables;
 import com.facebook.presto.metadata.FunctionRegistry;
-import com.facebook.presto.spi.InvocationConvention;
 import com.facebook.presto.spi.function.Convention;
 import com.facebook.presto.spi.function.FunctionDependency;
+import com.facebook.presto.spi.function.InvocationConvention;
 import com.facebook.presto.spi.function.OperatorDependency;
 import com.facebook.presto.spi.function.TypeParameter;
 import com.facebook.presto.spi.type.TypeManager;
