+++ presto-main/src/main/java/com/facebook/presto/operator/annotations/ScalarImplementationDependency.java
@@ -16,7 +16,7 @@ package com.facebook.presto.operator.annotations;
 import com.facebook.presto.metadata.BoundVariables;
 import com.facebook.presto.metadata.FunctionRegistry;
 import com.facebook.presto.metadata.Signature;
-import com.facebook.presto.spi.InvocationConvention;
+import com.facebook.presto.spi.function.InvocationConvention;
 import com.facebook.presto.spi.type.TypeManager;
 
 import java.lang.invoke.MethodHandle;
