+++ presto-spi/src/main/java/com/facebook/presto/spi/function/InvocationConvention.java
@@ -11,7 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.facebook.presto.spi;
+package com.facebook.presto.spi.function;
 
 import java.util.List;
